[FIX] web_view_leaflet_map: add Create button and fix navigate visibility
Restore functionality lost during rebase with upstream PR #434:

- Add Create/New button to map view control panel (from PR #434)
- Add onClickCreate() and createRecord() methods to controller
- Fix navigate button visibility (opacity 0 -> 0.6 when not hovered)

Clean up dead code from upstream refactor:

- Remove unused JS/XML files from views/leaflet_map/ directory
- Keep only leaflet_map_renderer.css which is used by our implementation
- Our custom implementation in leaflet_map_view/ has all features
