--- web_view_leaflet_map/static/src/views/leaflet_map/leaflet_map_renderer.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates xml:space="preserve">
-
-    <t t-name="web_view_leaflet_map.MapRenderer" owl="1">
-        <div class="o_leaflet_main_container h-100" t-ref="root">
-            <div t-ref="mapContainer" class="o_leaflet_map_container h-100" />
-        </div>
-    </t>
-
-</templates>
